Loading graph from file done and checking for memory leaks

+++ src/libgraph/graph.c
@@ -1,6 +1,10 @@
 #include "graph.h"
 #include "../liblist/list.h"
 
+bool is_node_oob(Graph *self, int nodeName){
+  return self->nbMaxNodes < nodeName;
+}
+
 //create a graph with the right number of nodes
 // error 1 : unexpected allocation error
 int create_graph(Graph *self, size_t maxNodes, bool isDirected){
@@ -18,9 +22,99 @@ int create_graph(Graph *self, size_t maxNodes, bool isDirected){
 
 //load a graph from a file
 int load_graph(Graph *self, const char *graphFile){
+#define BUFFER_SIZE 255
+  // Graph related vars
+  int nbMaxNodes = -1;
+  bool isDirected = false;
+
+  // File reading related
+  char buffer[BUFFER_SIZE];
+  char *edges[BUFFER_SIZE];
+  char cIsDirected;
   FILE* file = fopen(graphFile, "r");
-  // TODO: READ FILE AND STORE IT in MEMORY
+
+  // Skip the first commented line
+  fgets(buffer, BUFFER_SIZE, file);
+  // Reading the maximum number of nodes for our graph
+  fscanf(file, "%d", &nbMaxNodes);
+  // go to the next line
+  fgets(buffer, BUFFER_SIZE, file);
+
+  // Skip the commented line
+  fgets(buffer, BUFFER_SIZE, file);
+  // Reading the directed parameter
+  cIsDirected = fgetc(file);
+  if(cIsDirected == 'y'){
+    isDirected = true;
+  } else {
+    isDirected = false;
+  }
+  // We can now create our graph struct
+  create_graph(self, nbMaxNodes, isDirected);
+  // Go to next line
+  fgets(buffer, BUFFER_SIZE, file);
+
+  // Skip the commented line
+  fgets(buffer, BUFFER_SIZE, file);
+  // Reading graph nodes and edges
+  while(!feof(file)){
+    int nodeName;
+    // Scan for the nodeName
+    fscanf(file, "%d:", &nodeName);
+    // Decrease nodeName to get the real name in memory
+    nodeName--;
+    // Create the node in our graph
+    add_node(self, nodeName);
+    // Get edges for that node
+    fgets(buffer, BUFFER_SIZE, file);
+    // Storing edges informations in memory to compute it later
+    edges[nodeName] = (char*)malloc(strlen(buffer)+1);
+    memcpy(edges[nodeName], buffer, strlen(buffer)+1);
+    edges[nodeName][strlen(buffer)+1] = '\0';
+  }
+  // We have finish with our file so we can close it
   fclose(file);
+
+  // Compute graph edges
+  for(int i = 0; i < nbMaxNodes; i++) {
+    // First, check if our current node is existing
+    if(is_node_exists(self, i)){
+      // edges vars
+      int neighbourName;
+      int edgeName;
+      // scanning vars
+      char *edge = edges[i];
+      // Search for the first comma in edge list
+      char *firstComma = strchr(edge, ',');
+      // If there's a coma then we have multiple edges for this node
+      while(firstComma != NULL){
+        // looking for the size of string we have to allocate to get the edge description
+        int size = strcspn(edge, ",");
+        // create a temporary string to read the edge description
+        char *edgeDescription = (char*)malloc(size);
+        // copying the edge description in our temporary string
+        strncpy(edgeDescription, edge, size);
+        // read our edge description from our string
+        sscanf(edgeDescription, " (%d/%d)", &neighbourName, &edgeName);
+        // Now that we read the description into our vars we can free the edgeDescription
+        free(edgeDescription);
+        // creating the edge from our edge values
+        add_edge(self, i, neighbourName-1, edgeName, 0, true);
+        // then we can go to the next edge of the node (++firstComma to get the string after the first comma)
+        edge = ++firstComma;
+        // then searching for the next comma
+        firstComma = strchr(edge, ',');
+      }
+      // if there is no more comma and we still have an edge description, compute it
+      if(firstComma == NULL && strlen(edge) > 2){
+        sscanf(edge, " (%d/%d)", &neighbourName, &edgeName);
+        add_edge(self, i, neighbourName-1, edgeName, 0, true);
+      }
+      // We have finish to compute edges for this node, we free our temporary memory
+      free(edges[i]);
+    }
+  }
+
   return 0;
 }
 
@@ -28,7 +122,7 @@ int load_graph(Graph *self, const char *graphFile){
 // error 1 : node OOB
 // error 2 : unexpected allocation error
 int add_node(Graph *self, int nodeName){
-  if(nodeName < self->nbMaxNodes){
+  if(!is_node_oob(self, nodeName)){
     self->adjList[nodeName] = malloc(sizeof(Neighbour*));
     if(self->adjList[nodeName] == NULL){
       return 2;
@@ -68,10 +162,6 @@ int add_edge(Graph *self, int fromName, int toName, int edgeName, int Weight, bo
   }
 }
 
-bool is_node_oob(Graph *self, int nodeName){
-  return self->nbMaxNodes < nodeName;
-}
-
 // Check if a node already exists
 bool is_node_exists(Graph* self, int nodeName){
   return self->adjList[nodeName] != NULL;
@@ -99,29 +189,32 @@ int remove_node(Graph *self, int nodeName){
   if(!is_node_exists(self, nodeName)){
     return 10;
   }
-  // TODO: Supprimer une node
   // Supression des edge reliées à la node
   for(int i = 0; i < self->nbMaxNodes; i++){
     if(is_node_exists(self, i)){
-      //search and delete node from endpoint
+      deleteEdgeFromNodeName(self->adjList[i], nodeName);
     }
   }
   // Supression de la node en elle même
   destroyList(self->adjList[nodeName]);
-  self->adjList[nodeName] = NULL;
+  free(self->adjList[nodeName]);
   return 0;
 }
 
 //Delete an edge
 int remove_edge(Graph *self, int edgeName){
-  // TODO: Supprimer une edge
-  // Supprime l'edge
   for(int i = 0; i < self->nbMaxNodes; i++){
-    deleteEdge(self->adjList[i], edgeName);
+    if(is_node_exists(self, i)){
+      int error = deleteEdge(self->adjList[i], edgeName);
+      if (error != 0){
+        return error;
+      }
+    }
   }
   return 0;
 }
 
+// TODO: gérer les erreurs?
 int output_graph_to_stream(Graph *self, FILE *stream){
   fputs("# maximum number of node\n", stream);
   fprintf(stream, "%zu\n", self->nbMaxNodes);
@@ -135,7 +228,7 @@ int output_graph_to_stream(Graph *self, FILE *stream){
   for(int i = 0; i < self->nbMaxNodes; i++){
     if(is_node_exists(self, i)){
       fprintf(stream, "%d: ", i+1);
-      // call list output
+      outputList(self->adjList[i], stream);
       fputs("\n", stream);
     }
   }
@@ -152,12 +245,21 @@ int save_graph(Graph *self, const char *graphFile){
   FILE *file = fopen(graphFile, "w");
   int error = output_graph_to_stream(self, file);
   fclose(file);
-
   return error;
 }
 
 //Free graph and quit
 int quit(Graph *self){
-  //Save and free
+  if(self != NULL){
+    if(self->adjList != NULL){
+      for (size_t i = 0; i < self->nbMaxNodes; i++){
+        if(is_node_exists(self, i)){
+          destroyList(self->adjList[i]);
+          free(self->adjList[i]);
+        }
+      }
+      free(self->adjList);
+    }
+  }
   return 0;
 }
